Moved dates finalization to DateController.php instead of MeetingController.php
